rename SetInt to SetInt64

+++ go/bls/mcl.go
@@ -36,9 +36,9 @@ func (x *Fr) Clear() {
 }
 
 // SetInt --
-func (x *Fr) SetInt(v int) {
+func (x *Fr) SetInt64(v int64) {
 	// #nosec
-	C.mclBnFr_setInt(x.getPointer(), C.int(v))
+	C.mclBnFr_setInt(x.getPointer(), C.int64_t(v))
 }
 
 // SetString --
